Optimize rank/quality calculations

+++ obra_upgrade_calculator/rankings.py
@@ -3,12 +3,13 @@
 from __future__ import unicode_literals
 
 import logging
+from collections import defaultdict
 from datetime import date, timedelta
 
 from peewee import fn
 
 from .data import DISCIPLINE_MAP
-from .models import Event, Person, Points, Quality, Race, Rank, Result
+from .models import Event, Quality, Race, Rank, Result
 
 try:
     import ujson as json
@@ -18,49 +19,55 @@ except ImportError:
 logger = logging.getLogger(__name__)
 
 
-def get_rank(person, upgrade_discipline, end_date=None):
+def get_ranks(upgrade_discipline, end_date=None):
+    """
+    Return a dict of everyone's rank for this discipline as of a given date
+    """
     if not end_date:
         end_date = date.today()
     start_date = end_date - timedelta(days=365)
 
     ranks = [600] * 5
-    query = (Rank.select(fn.json_group_array(Rank.value, coerce=False))
+    query = (Rank.select(Result.person_id, fn.json_group_array(Rank.value, coerce=False))
                  .join(Result, src=Rank)
                  .join(Race, src=Result)
-                 .join(Event, src=Result)
-                 .where(Result.person_id == person.id)
-                 .where(Race.date >= start_date)
-                 .where(Race.date <= end_date)
-                 .where(Event.discipline << DISCIPLINE_MAP[upgrade_discipline])
-                 .order_by(Rank.value.asc())
-                 .limit(5))
-    ranks += json.loads(query.scalar())
-    return sum(sorted(ranks)[:5]) / 5
-
-
-def calculate_race_ranks(upgrade_discipline):
-    # Delete all Rank and Quality data for this discipline and recalc from scratch
-    # FIXME - add incremental support and make complete recalc selectable
-    (Rank.delete()
-         .where(Rank.result_id << (Rank.select(Rank.id)
-                                       .join(Result, src=Rank)
-                                       .join(Race, src=Result)
-                                       .join(Event, src=Race)
-                                       .where(Event.discipline << DISCIPLINE_MAP[upgrade_discipline])))
-         .execute())
-
-    (Quality.delete()
-            .where(Quality.result_id << (Quality.select(Quality.id)
-                                                .join(Race, src=Quality)
-                                                .join(Event, src=Race)
-                                                .where(Event.discipline << DISCIPLINE_MAP[upgrade_discipline])))
-            .execute())
-
-    races = (Race.select(Race, Event.name)
                  .join(Event, src=Race)
+                 .where(Race.date >= start_date)
+                 .where(Race.date < end_date)
+                 .where(Event.discipline << DISCIPLINE_MAP[upgrade_discipline])
+                 .group_by(Result.person_id))
+
+    logger.debug('Got {} People in {} between {} and {}'.format(query.count(), upgrade_discipline, start_date, end_date))
+    return defaultdict(lambda: 600, ((k, sum(sorted(ranks + json.loads(v))[:5]) / 5) for k, v in query.tuples()))
+
+
+def calculate_race_ranks(upgrade_discipline, incremental=False):
+    # Delete all Rank and Quality data for this discipline and recalc from scratch
+
+    if not incremental:
+        (Rank.delete()
+             .where(Rank.result_id << (Result.select(Result.id)
+                                             .join(Race, src=Result)
+                                             .join(Event, src=Race)
+                                             .where(Event.discipline << DISCIPLINE_MAP[upgrade_discipline])))
+             .execute())
+
+        (Quality.delete()
+                .where(Quality.race_id << (Race.select(Race.id)
+                                               .join(Event, src=Race)
+                                               .where(Event.discipline << DISCIPLINE_MAP[upgrade_discipline])))
+                .execute())
+
+    prev_race = Race()
+    races = (Race.select(Race, Event)
+                 .join(Event, src=Race)
+                 .where(Race.id.not_in(Quality.select(fn.DISTINCT(Race.id))
+                                              .join(Race, src=Quality)
+                                              .join(Event, src=Race)
+                                              .where(Event.discipline << DISCIPLINE_MAP[upgrade_discipline])))
                  .where(Event.discipline << DISCIPLINE_MAP[upgrade_discipline])
                  .where(Race.categories.length() != 0)  # | (Race.name ** '%single%'))  #<-- uncomment to include SS
-                 .order_by(Race.date.asc()))
+                 .order_by(Race.date.asc(), Race.created.asc()))
 
     for race in races:
         """
@@ -70,28 +77,34 @@ def calculate_race_ranks(upgrade_discipline):
         4. Store (((Average all ranked finishers) - (quality.value)) * 2) / (race.results.count() - 1) as quality.points_per_place
         5. For each result, store quality.value + ((result.place - 1) * quality.points_per_place) as rank.value
         """
-        logger.info('{} {}: {}'.format(race.date, race.event.name, race.name))
-        results = (race.results.select(Result.id,
-                                       Result.place,
-                                       Person.id)
-                               .join(Person)
-                               .where(~(Result.place ** ('DN%')))
-                               .where(Result.place.cast('integer') > 0)
-                               .order_by(Result.place.cast('integer').asc()))
+        logger.info('Processing Race: [{}]{}: [{}]{} on {}'.format(race.event.id, race.event.name, race.id, race.name, race.date))
 
-        if results.count() <= 2:
+        results = (race.results.select()
+                               .where(~(Result.place.contains('dns')))
+                               .where(~(Result.place.contains('dnf')))
+                               .order_by(Result.id.asc()))
+        finishers = results.count()
+
+        if finishers <= 2:
+            logger.debug('Insufficient finishers: {}'.format(finishers))
+            Quality.create(race=race, value=0, points_per_place=0)
             continue
 
+        # Bulk cache everyone's ranks for this date so we don't have to re-query them all one by one
+        if prev_race.date != race.date:
+            people = get_ranks(upgrade_discipline, race.date)
+
         ranks = [600] * 5
-        ranks += [get_rank(result.person, upgrade_discipline, race.date) for result in results.limit(10)]
+        ranks += [people[result.person_id] for result in results.limit(10)]
         min_rank = min(ranks)
         top_average = sum(sorted(ranks)[:5]) / 5
 
-        ranks = [get_rank(result.person, upgrade_discipline, race.date) for result in results]
+        ranks = [people[result.person_id] for result in results]
         all_average = sum(ranks) / len(ranks)
         value = (all_average if all_average < top_average and all_average > min_rank else top_average) * 0.9
-        per_place = ((all_average - value) * 2) / (results.count() - 1)
+        per_place = ((all_average - value) * 2) / (finishers - 1)
 
+        logger.info('\tStart/Finishers:  {}/{}'.format(race.starters, finishers))
         logger.info('\tAverage of top 5: {}'.format(top_average))
         logger.info('\tAverage of field: {}'.format(all_average))
         logger.info('\tBest top 10 rank: {}'.format(min_rank))
@@ -99,61 +112,11 @@ def calculate_race_ranks(upgrade_discipline):
         logger.info('\tPoints per Place: {}'.format(per_place))
         Quality.create(race=race, value=value, points_per_place=per_place)
 
-        for result in results:
-            rank = value + ((int(result.place) - 1) * per_place)
-            rank = int(rank) if rank <= 590 else 590
-            Rank.create(result=result, value=rank)
-            logger.debug('\t\t{}: {} - {}'.format(result.person.id, result.place, rank))
+        insert_ranks = []
+        for zplace, result in enumerate(results):
+            rank = value + (zplace * per_place)
+            rank = rank if rank <= 590 else 590
+            insert_ranks.append((result, rank))
 
-
-def dump_ranks(upgrade_discipline):
-    start_date = date.today() - timedelta(days=365)
-    people = (Person.select(Person)
-                    .join(Result, src=Person)
-                    .join(Race, src=Result)
-                    .join(Rank, src=Result)
-                    .where(Race.date >= start_date)
-                    .where(Event.discipline << DISCIPLINE_MAP[upgrade_discipline])
-                    .group_by(Person)
-                    .order_by(Person.last_name.collate('NOCASE').asc(),
-                              Person.first_name.collate('NOCASE').asc()))
-
-    ranks = [(person, get_rank(person, upgrade_discipline)) for person in people]
-    i = 0
-    last_rank = 0
-    print('Place| Cat   | Name                    : Rank points')
-    for (person, rank) in sorted(ranks, key=lambda r: r[1]):
-        if last_rank != rank:
-            last_rank = rank
-            i += 1
-        cat = (Points.select(Points.sum_categories)
-                     .join(Result, src=Points)
-                     .join(Race, src=Result)
-                     .join(Event, src=Race)
-                     .where(Result.person == person)
-                     .where(Event.discipline << DISCIPLINE_MAP[upgrade_discipline])
-                     .order_by(Race.date.desc())
-                     .limit(1)
-                     .scalar())
-        cat = '/'.join(str(c) for c in cat) if cat else 'SS'
-        name = '{}, {}'.format(person.last_name, person.first_name)
-        print('{0:<4} | {1:<5} | {2:<24}: {3:>3} points'.format(i, cat, name, rank))
-
-
-def dump_rank_history(person, upgrade_discipline):
-    start_date = date.today() - timedelta(days=365)
-    ranks = (Rank.select(Rank,
-                         Result,
-                         Person,
-                         Race)
-                 .join(Result, src=Rank)
-                 .join(Person, src=Result)
-                 .join(Race, src=Result)
-                 .join(Event, src=Race)
-                 .where(Result.person == person)
-                 .where(Race.date >= start_date)
-                 .where(Event.discipline << DISCIPLINE_MAP[upgrade_discipline])
-                 .order_by(Race.date.asc()))
-    print('{}, {}: {}'.format(person.last_name, person.first_name, get_rank(person, upgrade_discipline)))
-    for rank in ranks:
-        print('\t{0:>3} points | {1} - {2:<2} in {3}'.format(rank.value, rank.result.race.date, rank.result.place, rank.result.race.name))
+        Rank.insert_many(insert_ranks, fields=[Rank.result, Rank.value]).on_conflict_replace().execute()
+        prev_race = race
